+++ backend/src/main/java/heartbeat/client/dto/pipeline/buildkite/ProviderDTO.java
@@ -1,30 +1,24 @@
-package heartbeat.client.dto.pipeline.buildKite;
+package heartbeat.client.dto.pipeline.buildkite;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.Date;
 
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class CreatedByDTO {
+public class ProviderDTO {
 
 	private String id;
 
-	private String graphql_id;
+	private ProviderSettingsDTO settings;
 
-	private String name;
-
-	private String email;
-
-	private String avatar_url;
-
-	private Date created_at;
+	@JsonProperty("webhook_url")
+	private String webhookUrl;
 
 }
